Updating interface and model objects for a single database
