+++ dataactcore/models/baseInterface.py
@@ -4,16 +4,20 @@ from sqlalchemy.orm import sessionmaker, scoped_session
 from sqlalchemy.orm.exc import NoResultFound, MultipleResultsFound
 from dataactcore.utils.responseException import ResponseException
 from dataactcore.utils.statusCode import StatusCode
-
+from dataactcore.config import CONFIG_DB
 
 class BaseInterface(object):
     """ Abstract base interface to be inherited by interfaces for specific databases
     """
     #For Flask Apps use the context for locals
     IS_FLASK = True
-    dbName = None  # Should be overwritten by child classes
-    dbConfig = None  # Should be overwritten by child classes
+    dbConfig = None
     logFileName = "dbErrors.log"
+    dbName = dbConfig['db_name']
+    Session = None
+    engine = None
+    session = None
+    dbConfig = CONFIG_DB
 
     def __init__(self):
         if(self.session != None):
@@ -50,6 +54,15 @@ class BaseInterface(object):
         except (KeyError, AttributeError):
             # KeyError will occur in Python 3 on engine dispose
             pass
+
+    @staticmethod
+    def getDbName():
+        """Return database name."""
+        return BaseInterface.dbName
+
+    def getSession(self):
+        """ Return current active session """
+        return self.session
 
     @classmethod
     def getCredDict(cls):
