+++ dataactcore/models/validationModels.py
@@ -2,12 +2,7 @@
 
 from sqlalchemy import Column, Integer, Text, ForeignKey, Boolean
 from sqlalchemy.orm import relationship
-from dataactcore.models.validationBase import Base
-# the following lines import models that are maintained in
-# separate files but live in the validation database.
-# they're imported here so alembic will see them.
-import dataactcore.models.domainModels as domainModels
-import dataactcore.models.stagingModels as stagingModels
+from dataactcore.models.baseModel import Base
 
 class FileType(Base):
     __tablename__ = "file_type"
